setup util base matrix class
