+++ transformer_lens/FactoredMatrix.py
@@ -12,34 +12,14 @@ from typing import List, Tuple, Union, overload
 import torch
 from jaxtyping import Float
 
-import transformer_lens.utils as utils
+import transformer_lens.utilities.tensor_utils as tensor_utils
 
 
-class FactoredMatrix:
+
+class FactoredMatrix(tensor_utils.BaseTransformerLensMatrix):
     """
     Class to represent low rank factored matrices, where the matrix is represented as a product of two matrices. Has utilities for efficient calculation of eigenvalues, norm and SVD.
     """
-
-    def __init__(
-        self,
-        A: Float[torch.Tensor, "... ldim mdim"],
-        B: Float[torch.Tensor, "... mdim rdim"],
-    ):
-        self.A = A
-        self.B = B
-        assert self.A.size(-1) == self.B.size(
-            -2
-        ), f"Factored matrix must match on inner dimension, shapes were a: {self.A.shape}, b:{self.B.shape}"
-        self.ldim = self.A.size(-2)
-        self.rdim = self.B.size(-1)
-        self.mdim = self.B.size(-2)
-        self.has_leading_dims = (self.A.ndim > 2) or (self.B.ndim > 2)
-        self.shape = torch.broadcast_shapes(self.A.shape[:-2], self.B.shape[:-2]) + (
-            self.ldim,
-            self.rdim,
-        )
-        self.A = self.A.broadcast_to(self.shape[:-2] + (self.ldim, self.mdim))
-        self.B = self.B.broadcast_to(self.shape[:-2] + (self.mdim, self.rdim))
 
     @overload
     def __matmul__(
@@ -169,7 +149,7 @@ class FactoredMatrix:
         """
         Ua, Sa, Vha = torch.svd(self.A)
         Ub, Sb, Vhb = torch.svd(self.B)
-        middle = Sa[..., :, None] * utils.transpose(Vha) @ Ub * Sb[..., None, :]
+        middle = Sa[..., :, None] * tensor_utils.transpose(Vha) @ Ub * Sb[..., None, :]
         Um, Sm, Vhm = torch.svd(middle)
         U = Ua @ Um
         Vh = Vhb @ Vhm
@@ -239,11 +219,11 @@ class FactoredMatrix:
         """
         return FactoredMatrix(
             self.U * self.S.sqrt()[..., None, :],
-            self.S.sqrt()[..., :, None] * utils.transpose(self.Vh),
+            self.S.sqrt()[..., :, None] * tensor_utils.transpose(self.Vh),
         )
 
     def get_corner(self, k=3):
-        return utils.get_corner(self.A[..., :k, :] @ self.B[..., :, :k], k)
+        return tensor_utils.get_corner(self.A[..., :k, :] @ self.B[..., :, :k], k)
 
     @property
     def ndim(self) -> int:
@@ -253,7 +233,7 @@ class FactoredMatrix:
         """
         Collapses the left side of the factorization by removing the orthogonal factor (given by self.U). Returns a (..., mdim, rdim) tensor
         """
-        return self.S[..., :, None] * utils.transpose(self.Vh)
+        return self.S[..., :, None] * tensor_utils.transpose(self.Vh)
 
     def collapse_r(self) -> Float[torch.Tensor, "*leading_dims ldim mdim"]:
         """
