+++ transformer_lens/utilities/tensor_utils.py
@@ -12,14 +12,34 @@ import numpy as np
 import torch
 from jaxtyping import Float, Int
 
-from transformer_lens.FactoredMatrix import FactoredMatrix
+class BaseTransformerLensMatrix:
+    def __init__(
+        self,
+        A: Float[torch.Tensor, "... ldim mdim"],
+        B: Float[torch.Tensor, "... mdim rdim"],
+    ):
+        self.A = A
+        self.B = B
+        assert self.A.size(-1) == self.B.size(
+            -2
+        ), f"Factored matrix must match on inner dimension, shapes were a: {self.A.shape}, b:{self.B.shape}"
+        self.ldim = self.A.size(-2)
+        self.rdim = self.B.size(-1)
+        self.mdim = self.B.size(-2)
+        self.has_leading_dims = (self.A.ndim > 2) or (self.B.ndim > 2)
+        self.shape = torch.broadcast_shapes(self.A.shape[:-2], self.B.shape[:-2]) + (
+            self.ldim,
+            self.rdim,
+        )
+        self.A = self.A.broadcast_to(self.shape[:-2] + (self.ldim, self.mdim))
+        self.B = self.B.broadcast_to(self.shape[:-2] + (self.mdim, self.rdim))
 
 
 def get_corner(tensor, n=3):
     # Prints the top left corner of the tensor
     if isinstance(tensor, torch.Tensor):
         return tensor[tuple(slice(n) for _ in range(tensor.ndim))]
-    elif isinstance(tensor, FactoredMatrix):
+    elif isinstance(tensor, BaseTransformerLensMatrix):
         return tensor[tuple(slice(n) for _ in range(tensor.ndim))].AB
 
 
@@ -114,7 +134,7 @@ def check_structure(t1: torch.Tensor, t2: torch.Tensor, *, verbose: bool = False
 
 
 def composition_scores(
-    left: "FactoredMatrix", right: "FactoredMatrix", broadcast_dims=True
+    left: BaseTransformerLensMatrix, right: BaseTransformerLensMatrix, broadcast_dims=True
 ) -> Union[
     Float[torch.Tensor, "*leading_dims"],
     Float[torch.Tensor, "*leading_dims_left_and_right"],
